# backend/docs/swagger.yaml
basePath: /v1/api/radar-hub-manager
definitions:
  github_com_lehaisonagentai2_radar-hub-manager_backend_internal_models.Command:
    properties:
      acknowledged_at:
        type: integer
      content:
        type: string
      created_at:
        type: integer
      from_user_id:
        description: UUID dưới dạng chuỗi
        type: string
      id:
        type: integer
      sent_at:
        type: integer
      to_station_id:
        type: integer
    type: object
  github_com_lehaisonagentai2_radar-hub-manager_backend_internal_models.RoleName:
    enum:
    - ADMIN
    - OPERATOR
    - HQ
    type: string
    x-enum-comments:
      RoleAdmin: Quản trị hệ thống
      RoleHQ: Cán bộ Sở chỉ huy
      RoleOperator: Nhân viên vận hành trạm
    x-enum-descriptions:
    - Quản trị hệ thống
    - Nhân viên vận hành trạm
    - Cán bộ Sở chỉ huy
    x-enum-varnames:
    - RoleAdmin
    - RoleOperator
    - RoleHQ
...
  github_com_lehaisonagentai2_radar-hub-manager_backend_internal_models.Station:
    properties:
      created_at:
        type: integer
      distance_to_coast:
        description: km
        type: number
      elevation:
        description: mét so với mực nước biển
        type: number
      id:
        type: integer
      latitude:
        type: number
      longitude:
        type: number
      name:
        type: string
      note:
        description: ghi chú tự do
        type: string
      status:
        description: '"ACTIVE" / "INACTIVE" (tính từ lịch)'
        type: string
      updated_at:
        type: integer
    type: object
  github_com_lehaisonagentai2_radar-hub-manager_backend_internal_models.User:
    properties:
      created_at:
        type: integer
      full_name:
        type: string
      id:
        description: UUID dưới dạng chuỗi
        type: integer
      last_login:
        type: integer
      password:
        description: Chỉ dùng khi tạo mới, không trả về
        type: string
      role_id:
        $ref: '#/definitions/github_com_lehaisonagentai2_radar-hub-manager_backend_internal_models.RoleName'
      station:
        $ref: '#/definitions/github_com_lehaisonagentai2_radar-hub-manager_backend_internal_models.Station'
      station_id:
        description: 'Nếu là OPERATOR: trạm mà người dùng trực thuộc'
        type: integer
      updated_at:
        type: integer
      username:
        type: string
    type: object
  internal_handlers.CreateCommandRequest:
    properties:
      content:
        type: string
      to_station_id:
        type: integer
    required:
    - content
    - to_station_id
    type: object
...
  internal_handlers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  internal_handlers.LoginRequest:
    properties:
      password:
        example: password123
        type: string
      username:
        example: admin
        type: string
    required:
    - password
    - username
    type: object
  internal_handlers.LoginResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/github_com_lehaisonagentai2_radar-hub-manager_backend_internal_models.User'
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the API server for the Radar Hub Manager system
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Radar Hub Manager API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return user info with JWT token
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successful login
          schema:
            $ref: '#/definitions/internal_handlers.LoginResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      summary: User login
      tags:
      - auth
  /auth/me:
    get:
      consumes:
      - application/json
      description: Get information about the currently authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: User information
          schema:
            $ref: '#/definitions/github_com_lehaisonagentai2_radar-hub-manager_backend_internal_models.User'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get current user info
      tags:
      - auth
  /commands:
    get:
      description: List all commands or commands for a specific station
      parameters:
      - description: Station ID to filter commands
        in: query
        name: station_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_lehaisonagentai2_radar-hub-manager_backend_internal_models.Command'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List commands
      tags:
      - commands
    post:
      consumes:
      - application/json
      description: Create a new command to send to a station (HQ only)
      parameters:
      - description: Command data
        in: body
        name: command
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.CreateCommandRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_lehaisonagentai2_radar-hub-manager_backend_internal_models.Command'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new command
      tags:
      - commands
  /commands/{id}:
    get:
      description: Get a specific command by ID
      parameters:
      - description: Command ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lehaisonagentai2_radar-hub-manager_backend_internal_models.Command'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a command by ID
      tags:
      - commands
  /commands/{id}/acknowledge:
    put:
      consumes:
      - application/json
      description: Acknowledge a command by ID (Operator only)
      parameters:
      - description: Command ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lehaisonagentai2_radar-hub-manager_backend_internal_models.Command'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Acknowledge a command
      tags:
      - commands
  /station-commands/station/{station_id}/unacknowledged:
    get:
      description: List unacknowledged commands for a specific station (Operator only)
      parameters:
      - description: Station ID
        in: path
        name: station_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_lehaisonagentai2_radar-hub-manager_backend_internal_models.Command'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List unacknowledged commands
      tags:
      - commands
...
securityDefinitions:
  ApiKeyAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"